requirement.txt
pypdfium2==4.23.0           # PDF to image conversion
pytesseract==0.3.10          # OCR text extraction
Pillow==10.0.1               # Image processing
PyPDF2==3.0.1                # PDF text extraction
langchain==0.1.5              # LLM inference
jsonformer==0.1.6             # JSON generation from LLM
requests==2.31.0              # API calls for pending PO report
